Add explicit type conversions where necessary: device/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2947537
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/master@{#891362}

+++ device/fido/win/type_conversions_unittest.cc
@@ -5,6 +5,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "device/fido/win/type_conversions.h"
 
+#include "base/numerics/safe_conversions.h"
 #include "base/strings/utf_string_conversions.h"
 #include "components/cbor/values.h"
 #include "components/cbor/writer.h"
@@ -94,9 +95,9 @@ TEST(TypeConversionsTest, ToAuthenticatorMakeCredentialResponse) {
         ToAuthenticatorMakeCredentialResponse(WEBAUTHN_CREDENTIAL_ATTESTATION{
             WEBAUTHN_CREDENTIAL_ATTESTATION_VERSION_3,
             test.format,
-            test.authenticator_data.size(),
+            base::checked_cast<DWORD>(test.authenticator_data.size()),
             const_cast<unsigned char*>(test.authenticator_data.data()),
-            test.cbor_attestation_statement.size(),
+            base::checked_cast<DWORD>(test.cbor_attestation_statement.size()),
             const_cast<unsigned char*>(test.cbor_attestation_statement.data()),
             // dwAttestationDecodeType and pvAttestationDecode are ignored.
             WEBAUTHN_ATTESTATION_DECODE_NONE,
